+++ src/lib/utils/eventSignOut.js
@@ -1,8 +1,10 @@
 import { auth, logOut } from '../../firebase/authenticationFirebase.js';
 
-const signOut = () => {
-  const result = logOut(auth);
-      result.then(() => {
+const eventButtonSignOut = () => {
+  const btnExit = document.getElementById('exitButtonId');
+  if (btnExit) {
+    btnExit.addEventListener('click', () => {
+      logOut(auth).then(() => {
         const getUserCredential = localStorage.getItem('UserCredentialFb');
         console.log(getUserCredential, 'Has cerrado sesión');
         localStorage.removeItem('UserCredentialFb');
@@ -12,13 +14,6 @@ const signOut = () => {
       }).catch((error) => {
         console.error(error);
       });
-}
-
-const eventButtonSignOut = () => {
-  const btnExit = document.getElementById('exitButtonId');
-  if (btnExit) {
-    btnExit.addEventListener('click', () => {
-      signOut();
     });
   }
 };
